Remove lib/ezid/helper.rb
In previous commit, added this file from the orignal hypatia-new
code. However, instead of porting the whole file and then
customizing it, will probably take snippets of the code instead.
One reason: A lot of the original code is hypatia specific, and
will need to be rewritten for hyacinth. Also, lots of non-rubocop
compliant code.